refactor: 영속화 관련 기능 패키지 변경
