+++ space-d/src/main/java/com/dnd/spaced/core/quiz/infrastructure/persistence/QuizMetadataGatewayRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.quiz.infrastructure.repository;
+package com.dnd.spaced.core.quiz.infrastructure.persistence;
 
 import static com.dnd.spaced.core.quiz.domain.QQuizMetadata.quizMetadata;
