+++ space-d/src/main/java/com/dnd/spaced/core/quiz/infrastructure/persistence/QuizQuestionCrudRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.quiz.infrastructure.repository;
+package com.dnd.spaced.core.quiz.infrastructure.persistence;
 
 import com.dnd.spaced.core.quiz.domain.QuizQuestion;
 import java.util.List;
