+++ space-d/src/main/java/com/dnd/spaced/core/quiz/infrastructure/persistence/TodayQuizCrudRepository.java
@@ -1,4 +1,4 @@
-package com.dnd.spaced.core.quiz.infrastructure.repository;
+package com.dnd.spaced.core.quiz.infrastructure.persistence;
 
 import com.dnd.spaced.core.quiz.domain.TodayQuiz;
 import org.springframework.data.repository.CrudRepository;
